adding test for ccxt methods depending on exchange instantiated.
